# ai-process-mining/docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "AI Process Mining Agent",
  "description": "Event logs \u2192 process discovery \u2192 metrics \u2192 bottlenecks \u2192 deviations \u2192 variants \u2192 AI explanation. The deterministic engine works without an LLM; the AI layer only explains computed metrics.",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/process/{dataset_id}/overview": {
   "get": {
    "tags": [
     "process"
    ],
    "summary": "Overview",
    "operationId": "overview_api_process__dataset_id__overview_get",
    "parameters": [
     {
      "name": "dataset_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Dataset Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/process/{dataset_id}/graph": {
   "get": {
    "tags": [
     "process"
    ],
    "summary": "Graph",
    "operationId": "graph_api_process__dataset_id__graph_get",
    "parameters": [
     {
      "name": "dataset_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Dataset Id"
      }
     },
     {
      "name": "min_case_pct",
      "in": "query",
      "required": false,
      "schema": {
       "type": "number",
       "maximum": 100.0,
       "minimum": 0.0,
       "default": 0,
       "title": "Min Case Pct"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/process/{dataset_id}/variants": {
   "get": {
    "tags": [
     "process"
    ],
    "summary": "Variants",
    "operationId": "variants_api_process__dataset_id__variants_get",
    "parameters": [
     {
      "name": "dataset_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Dataset Id"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "default": 0,
       "title": "Offset"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/process/{dataset_id}/variants/{variant_id}/cases": {
   "get": {
    "tags": [
     "process"
    ],
    "summary": "Variant Cases",
    "operationId": "variant_cases_api_process__dataset_id__variants__variant_id__cases_get",
    "parameters": [
     {
      "name": "dataset_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Dataset Id"
      }
     },
     {
      "name": "variant_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Variant Id"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "default": 0,
       "title": "Offset"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 1000,
       "minimum": 1,
       "default": 100,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/process/{dataset_id}/bottlenecks": {
   "get": {
    "tags": [
     "process"
    ],
    "summary": "Bottlenecks",
    "description": "Bottleneck candidates. Pass w_* query params to re-weight the score transparently.",
    "operationId": "bottlenecks_api_process__dataset_id__bottlenecks_get",
    "parameters": [
     {
      "name": "dataset_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Dataset Id"
      }
     },
     {
      "name": "w_median_wait",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "number"
        },
        {
         "type": "null"
        }
       ],
       "title": "W Median Wait"
      }
     },
     {
      "name": "w_p90_wait",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "number"
        },
        {
         "type": "null"
        }
       ],
       "title": "W P90 Wait"
      }
     },
     {
      "name": "w_processing",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "number"
        },
        {
         "type": "null"
        }
       ],
       "title": "W Processing"
      }
     },
     {
      "name": "w_case_coverage",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "number"
        },
        {
         "type": "null"
        }
       ],
       "title": "W Case Coverage"
      }
     },
     {
      "name": "w_wait_share",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "number"
        },
        {
         "type": "null"
        }
       ],
       "title": "W Wait Share"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/process/{dataset_id}/deviations": {
   "get": {
    "tags": [
     "process"
    ],
    "summary": "Deviations",
    "operationId": "deviations_api_process__dataset_id__deviations_get",
    "parameters": [
     {
      "name": "dataset_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Dataset Id"
      }
     },
     {
      "name": "type",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Type"
      }
     },
     {
      "name": "activity",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Activity"
      }
     },
     {
      "name": "case_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Case Id"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "default": 0,
       "title": "Offset"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 1000,
       "minimum": 1,
       "default": 100,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/process/{dataset_id}/resources": {
   "get": {
    "tags": [
     "process"
    ],
    "summary": "Resources",
    "operationId": "resources_api_process__dataset_id__resources_get",
    "parameters": [
     {
      "name": "dataset_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Dataset Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/research/{dataset_id}/analyze": {
   "post": {
    "tags": [
     "ai"
    ],
    "summary": "Analyze",
    "operationId": "analyze_api_research__dataset_id__analyze_post",
    "parameters": [
     {
      "name": "dataset_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Dataset Id"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "anyOf": [
         {
          "$ref": "#/components/schemas/AnalyzeRequest"
         },
         {
          "type": "null"
         }
        ],
        "title": "Body"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/research/{dataset_id}/latest": {
   "get": {
    "tags": [
     "ai"
    ],
    "summary": "Latest",
    "operationId": "latest_api_research__dataset_id__latest_get",
    "parameters": [
     {
      "name": "dataset_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Dataset Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/research/{dataset_id}/report.md": {
   "get": {
    "tags": [
     "ai"
    ],
    "summary": "Latest Markdown",
    "operationId": "latest_markdown_api_research__dataset_id__report_md_get",
    "parameters": [
     {
      "name": "dataset_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Dataset Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/research/{dataset_id}/runs": {
   "get": {
    "tags": [
     "ai"
    ],
    "summary": "Runs",
    "operationId": "runs_api_research__dataset_id__runs_get",
    "parameters": [
     {
      "name": "dataset_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Dataset Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/query/{dataset_id}": {
   "post": {
    "tags": [
     "ai"
    ],
    "summary": "Query",
    "operationId": "query_api_query__dataset_id__post",
    "parameters": [
     {
      "name": "dataset_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Dataset Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/QueryRequest"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/QueryResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  }
 },
 "components": {
  "schemas": {
   "AnalyzeRequest": {
    "properties": {
     "focus": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 500
       },
       {
        "type": "null"
       }
      ],
      "title": "Focus",
      "description": "Optional area to emphasise."
     }
    },
    "type": "object",
    "title": "AnalyzeRequest"
   },
...
   "ChatTurn": {
    "properties": {
     "role": {
      "type": "string",
      "pattern": "^(user|assistant)$",
      "title": "Role"
     },
     "content": {
      "type": "string",
      "maxLength": 4000,
      "title": "Content"
     }
    },
    "type": "object",
    "required": [
     "role",
     "content"
    ],
    "title": "ChatTurn"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "QueryRequest": {
    "properties": {
     "question": {
      "type": "string",
      "maxLength": 1000,
      "minLength": 2,
      "title": "Question"
     },
     "history": {
      "items": {
       "$ref": "#/components/schemas/ChatTurn"
      },
      "type": "array",
      "maxItems": 12,
      "title": "History"
     }
    },
    "type": "object",
    "required": [
     "question"
    ],
    "title": "QueryRequest"
   },
   "QueryResponse": {
    "properties": {
     "answer": {
      "type": "string",
      "title": "Answer"
     },
     "evidence": {
      "items": {
       "type": "object"
      },
      "type": "array",
      "title": "Evidence"
     },
     "tools_used": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Tools Used"
     },
     "mode": {
      "type": "string",
      "title": "Mode"
     },
     "flags": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Flags",
      "default": []
     },
     "prompt_version": {
      "type": "string",
      "title": "Prompt Version"
     },
     "run_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Run Id"
     }
    },
    "type": "object",
    "required": [
     "answer",
     "evidence",
     "tools_used",
     "mode",
     "prompt_version"
    ],
    "title": "QueryResponse"
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}